% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotationF.R
\name{annotateInteractionsToGenes}
\alias{annotateInteractionsToGenes}
\title{Annotate interaction set to genes reported in geneAnnotation}
\usage{
annotateInteractionsToGenes(TSSextended, interactions)
}
\description{
Help f() to annotate loc2 | loc1 to genes with which they potentially 
interact
}
\author{
Inga Patarcic
}
\keyword{internal}
