% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotationF.R
\name{annotateToPromoter}
\alias{annotateToPromoter}
\title{Annotate genomic location(windows) to genes reported in geneAnnotation}
\usage{
annotateToPromoter(windows, TSSextended, geneAnnotations, identified)
}
\description{
Help f() to annotate loc1 (windows) to genes with which they potentially 
interact
}
\author{
Inga Patarcic
}
\keyword{internal}
