% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associateReg2Gene.R
\name{associateReg2Gene}
\alias{associateReg2Gene}
\title{Associate regulatory regions to genes}
\usage{
associateReg2Gene(input, method = "pearson", tag = NULL, scale = TRUE,
  cores = 1, B = 1000, ...)
}
\arguments{
\item{input}{a GRangesList that contains regulatory activity and
gene expression results. It has to have specific columns, see the
output of \code{\link{regActivityAroundTSS}} function.}

\item{method}{one of "pearson","spearman","dcor","elasticnet",
"randomForest". Default: "pearson". See Details for more.}

\item{tag}{a string to be used to append to results from methods. This string
will be attached to "n","coefs","pval" columns. Default NULL.}

\item{scale}{if TRUE (default) the the values for gene expression and
regulatory activity will be scaled to have 0 mean and unit variance using
\code{\link[base]{scale}} function.}

\item{cores}{number of cores to be used}

\item{B}{number of randomizations, default 1000. This procedure
is used to estimate
P-values for coeficients returned by different methods.}

\item{...}{further arguments to methods, not implemented yet}
}
\value{
a GRanges object containing every potential association and
between a regulatory region and TSS, and the estimated association statistic
, its P-value and Q-value.
}
\description{
The function associates/links regulatory regions to genes based
on regulatory activity and gene expression relationship.
}
\details{
The function implements four methods to associate regulatory
activity to genes: correlation (pearson or spearman), distance
correlation ("dcor"), elasticnet and random forests. Simple linear regression on
scaled data is equivalent to Pearson correlation. Distance correlation is
a metric for statistical dependence between two variables. It can capture
non-linear relationships different from correlation.
Elasticnet and randomForests are methods that
can capture additivity between regulatory activities and their relationship
to gene expression. For all the methods, P-values are estimated by shuffling
the gene expression vector \code{B} times and getting a null distribution
for the estimated statistic, and comparing the original statistic to the null
distribution. Estimated statistics are
correlation coefficients for correlation and distance correlation methods,
Regression coefficients for elasticnet and gini importance scores for random
forests.
}
\examples{
\dontrun{
input=readRDS("data/sampleRegActivityAroundTSS.rds")
i2=readRDS("data/Roadmap_H3K27ac.trial.rds")
lapply(i2,function(x) x[,])
i3=i2[1:5]
i3=i2[567:569]
b=associateReg2Gene(i3)

}

}
\author{
Altuna Akalin
}
