% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voteAssociationsGI.R
\name{assocTable2}
\alias{assocTable2}
\title{this function creates a table of keys and pvals from
a GRangesList which contains GRanges as the
output of associateReg2Gene or metaAssociate functions.
this is needed to process the interactions that
exist in many datasets and merge them into the same table with
stats from the statistical tests}
\usage{
assocTable2(grlist, qval, cutoff.column)
}
\arguments{
\item{grlist}{}

\item{qval}{q-value cutoff that will be used to filter elements in the
input \code{GRangesList}. This argument will only take effect when the
\code{GRanges} object has a numeric "qval" column. If the input object
lacks this column, every association in the object will be treated as a valid
association prediction.}

\item{cutoff.column}{(character,"pval" is defalut). 
Which statistics to filter:"qval" or "pval"}
}
\description{
this function creates a table of keys and pvals from
a GRangesList which contains GRanges as the
output of associateReg2Gene or metaAssociate functions.
this is needed to process the interactions that
exist in many datasets and merge them into the same table with
stats from the statistical tests
}
\author{
Altuna Akalin
}
\keyword{internal}
