% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkScriptsGRAdjusted.R
\name{benchmark}
\alias{benchmark}
\title{Benchmarks reg2gene models using benchmark data}
\usage{
benchmark(reg2Gene, benchReg, regCol = "reg", binary = FALSE)
}
\arguments{
\item{reg2Gene}{a GRanges object output (from \code{associatereg2Gene}).
GRanges object should have at least one meta-data column which stores
the second GRanges object. Usually,one GRanges object corresponds to the gene 
locations, preferably TSS location, whereas the other GRanges object 
corresponds to the regulatory region locations. Meta-data column name should
be "reg".}

\item{benchReg}{a GRanges object with at least one meta-data column which 
stores the second GRanges object named "reg".}

\item{regCol}{character (default "reg"), a column name of meta-data object.
Indicates the column where the location of the 2nd pair of GRanges is stored,
(usually regulatory region locations)}

\item{binary}{(def:FALSE) how many times reg2Gene interactions is observed in
the benchmark dataset. If TRUE, reports if overlap with benchmark dataset is
observed at least once).}
}
\value{
GRanges object which is equall to the reg2Gene input object except 
that a meta-data column reporting how many times interaction was overlapped
or whether reported interaction was overlapped is added and 
named "BenchmarkO"
}
\description{
any other modelling procedure implemente in \code{reg2gene} package,
and predefined benchmark dataset as GRanges object.
This function adds a metadata column with info about benchmarking success -
whether tested regions are benchmarked or not.  
By default it reportes how many times interactions is observed in the 
benchmark dataset. If binary is set to TRUE, then logical vector of TRUE 
(overlapping benchmark dataset at least once) and FALSE (not overlapping
benchmark dataset at all) is reported.
}
\details{
benchmark dataset is inputed are overlaped. ComplexOverlap between reg2Gene 
and benchmark object is performed, and for each input pair it is reported 
whether this pair is benchmarked or not,and how many times (if binary=F).
ComplexOverlap means that criss-cross overlap of interacting regions is 
performed; eg four input regions are compared: two interacting regions
from one dataset with two interacting regions from the other dataset
in the criss-cross manner. Thus, the input orientation of the 
interacting regions is not strict: the first pair of the region1
from one dataset will be compared to both; the first pair and the
second pair of region2(and its pair is compared in the opposite 
direction).
}
\examples{
benchmark(GRReg1_toy,GRReg2_toy)
benchmark(GRReg1_toy,GRReg2_toy,binary=TRUE)
}
\author{
Inga Patarcic
}
