% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BenchMarkedReg2Gene_toy}
\alias{BenchMarkedReg2Gene_toy}
\title{Sample GRanges to test confusionmatrix function}
\format{GRanges object with 6 ranges and 5 metadata columns}
\usage{
BenchMarkedReg2Gene_toy
}
\description{
Sample GRanges to test confusionmatrix function
}
\keyword{datasets}
