% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkGInteractions.R
\name{benchmarkGI}
\alias{benchmarkGI}
\title{Benchmarks reg2gene models using benchmark data}
\usage{
benchmarkGI(reg2Gene, benchData, binary = FALSE, nCores = 1,
  ignore.strand = FALSE, ...)
}
\arguments{
\item{reg2Gene}{a \code{\link[InteractionSet]{GInteractions}} object output
from \code{\link{associateReg2Gene}}). Usually, 1st GRanges object, or anchor1
corresponds to the enahncer location, whereas the other GRanges object 
corresponds to the regulatory region locations.}

\item{benchData}{a \code{\link[InteractionSet]{GInteractions}} object output
from \code{\link{associateReg2Gene}}) or a list of GInteractions object. 
Both regions are used in the benchmarking procedure. 
This object stores benchmarking informations eg interacting region
coordinates from techniques such as HiC,eQTL studies...}

\item{binary}{(def:FALSE) how many times reg2Gene interactions is observed in
the benchmark dataset(s). If TRUE, reports if overlap with benchmark dataset 
is observed at least once).}

\item{nCores}{possible to be runned in parallel. Argument for mclapply f();
how many cores to use.}

\item{ignore.strand}{argument to be passed to
\code{\link[IRanges]{findOverlaps}}. When set to TRUE, the strand 
information is ignored in the overlap analysis.}

\item{...}{further arguments to methods, not implemented yet}
}
\value{
GInteractions object with added benchmark results metadata [Bench 
column].Each column metadata column corresponds to one benchmark dataset 
analyzed if input is list() 
Values can be either 0/1 (not/benchmarked) or 0-n (how many times each 
gene-enhancer pair is benchmared).
}
\description{
The function that takes as input results of \code{\link{associateReg2Gene}} 
or any other modelling procedure implemente in \code{reg2gene} package,
and predefined benchmark dataset as 
\code{\link[InteractionSet]{GInteractions}} object.
This function adds a metadata column with info about benchmarking success -
whether tested regions are benchmarked or not.  
By default it reportes how many times interactions is observed in the 
benchmark dataset. If binary is set to TRUE, then vector of 0' and 1's is 
reported (1 - overlapping benchmark dataset at least once) and 0 (not
overlapping benchmark dataset at all).
}
\details{
GInteractions objects - an output of \code{\link{associateReg2Gene}}
[or a list of such objects] and benchmark dataset are overlaped. 
\code{\link[InteractionSet]{linkOverlaps}} between reg2Gene 
and benchmark object is performed, and for each input pair it is reported 
whether this pair is benchmarked or not,and how many times (if binary=F).
Criss-cross overlap of interacting regions is performed; thus is anchor1 from
benchmark dataset is overlapping anchor2 from tested dataset, than anchor2
from benchmark dataset needs to overlap anchor1 from tested datased, or
vice-versa.
}
\examples{
# Creating testing and benchmarking dataset
require(GenomicRanges)
require(InteractionSet)
   
   reg2Gene <- GInteractions(GRReg1_toy,GRReg1_toy$reg)
   benchData <- GInteractions(GRReg2_toy,GRReg2_toy$reg)
   
benchmarkGI(reg2Gene,
            benchData,
            binary=FALSE)
            
benchmarkGI(reg2Gene,
            benchData,
            binary=TRUE)             

   
##################   
# example for list:

# NOTE: anchor1.Bench1Exp & anchor1.Bench2Exp are expected/precalculated
# values for this benchmark example
 
  benchDataList <- list(benchData,reg2Gene)
  names(benchDataList) <- c("benchData1","benchData2")


reg2GeneB <- benchmarkGI(reg2Gene,
                        benchDataList,
                        ignore.strand=TRUE,
                        binary=FALSE,
                        nCores = 1)               
            
 # Checking what happends when anchor1&anchor2 both overlap only one region
 # in benchmark dataset?  OK, They are not benchmarked...
 
   benchmarkGI(reg2Gene[1],
               benchData[1])      

 # WARNING! 
 # However, one need to be careful when benchmarking anchors that overlap 
 # within test set (eg enhancer overlaps gene region), 
 # because these regions will be benchmarked.
   
   benchmarkGI(reg2Gene[5],
               benchData)           
                                    
}
\author{
Inga Patarcic
}
