% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkGInteractions.R
\name{benchmarkGIsimple}
\alias{benchmarkGIsimple}
\title{Benchmarks help function}
\usage{
benchmarkGIsimple(reg2Gene, benchData, ignore.strand, binary)
}
\description{
Benchmark reg2gene models using benchmark data but for only one GInteraction
object [not lists]
}
\details{
Read description for \code{benchmarkGI}
}
\author{
Inga Patarcic
}
\keyword{internal}
