% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkGInteractions.R
\name{benchmarkInteractionssimple}
\alias{benchmarkInteractionssimple}
\title{Benchmarks help function}
\usage{
benchmarkInteractionssimple(interactions, benchInteractions, ignore.strand,
  binary)
}
\description{
Benchmark interactions models using benchmark data but for only one 
GInteraction object [not lists]
}
\details{
Read description for \code{benchmarkInteractions}
}
\author{
Inga Patarcic
}
\keyword{internal}
