% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkScriptsGRAdjusted.R
\name{BenchMarkReg2Gene}
\alias{BenchMarkReg2Gene}
\title{Benchmarks Reg2gene models using Benchmark data}
\usage{
BenchMarkReg2Gene(Reg2Gene, Benchmark, binary = FALSE)
}
\arguments{
\item{Reg2Gene}{a GRanges object output from \code{linkReg2Gene} or \code{MetaReg2Gene}}

\item{Benchmark}{a GRanges object}

\item{binary}{(def:FALSE) how many times Reg2Gene interactions is
observed in the benchmark dataset. If TRUE, reports if overlap with
benchmark dataset is observed at least once).}
}
\value{
a GRanges object which corresponds to Reg2Gene object
but with column added which reports how many times interaction was overlapped
or whether reported interaction was overlapped
}
\description{
any other modelling procedure implemente in \code{reg2gene} package,
For input Reg2Gene object it adds a column with info whether reported interactions is 
benchmarked or not. By default it reportes how many times Reg2Gene interactions is
observed in the benchmark dataset. If binary is set to TRUE, then only logical vector of
TRUE (overlapping benchmark dataset at least once) and FALSE(not overlapping benchmark 
dataset at all) entries is reported.
}
\details{
GRanges object or dataframe as an output of linkReg2Gene or Reg2GeneMA 
and Benchmark dataset is inputed. ComplexOverlap between Reg2Gene and Benchmark 
object is performed, and for each input pair it is reported whether this pair is
benchmarked or not,and how many times (if binary=F)
}
