% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotF.R
\name{benchPlotHelp}
\alias{benchPlotHelp}
\title{Plots enhancer-promoter interactions}
\usage{
benchPlotHelp(name, benchInteractions, PlotRange, chr)
}
\description{
Help function to create benckmark tracks, especilly needed for a list of 
benchmark inputs.
}
\author{
IngaPa
}
\keyword{internal}
