% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateDisease.R
\name{compareReg2Traits}
\alias{compareReg2Traits}
\title{Compare two objects of genomic region~gene~trait associations}
\usage{
compareReg2Traits(reg2Trait, reg2Trait2, traitID = c("Trait", "Trait"))
}
\arguments{
\item{reg2Trait}{A GInteractions object, likely produced by 
[\code{\link{annotateRegionToTrait}}]. This object stores info about
genomic region~gene~trait associations, as follows: anchor1 corresponds
to genomic region (likely SNP location), anchor2 corresponds to gene 
location, and gene symbol and associated disease is stored as a meta-data.}

\item{reg2Trait2}{A GInteractions object, likely produced by 
[\code{\link{annotateRegionToTrait}}]. This object stores info about
genomic region~gene~trait associations, as follows: anchor1 corresponds
to genomic region (likely SNP location), anchor2 corresponds to gene}

\item{traitID}{character vector of length 2 which stores information
about column names where trait info are stored for both input objects. 
By default c("Trait","Trait")}
}
\value{
a dataframe with rows corresponding to traits, and two columns with
info about gene counts for these traits. Each column gathers info from one 
reg2Trait GInteractions object (variant~gene interactions + trait 
association stored as meta-data).
}
\description{
This function reports per trait number of reported 
genomic region~gene~trait from two different objects.
}
\details{
For each trait that is present in at least one of the reg2Trait 
GInteractions objects report how many associated genes is present in the 
first GInteractions object, and how many is present in the 2nd 
GInteractions object. It allows an easy identification of traits for which
[\code{\link{annotateGenomicRegions}}] analysis improved number of disease
associated genes.
}
\examples{
#[TO do]

}
\author{
Inga Patarcic
}
\keyword{internal}
