% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkScriptsGRAdjusted.R
\name{complexOverlaps}
\alias{complexOverlaps}
\title{Criss-cross overlap of interacting regions}
\usage{
complexOverlaps(reg1, reg2, reg.col)
}
\arguments{
\item{reg1}{- GRanges object with at least one meta-data column which stores
the second GRanges object. Usually,one GRanges object corresponds to the gene 
locations, preferably TSS location, whereas the other GRanges object 
corresponds to the regulatory region locations. Meta-data column name should
be "reg".}

\item{reg2}{- GRanges object with at least one meta-data column which stores
the second GRanges object.  Meta-data column name should be "reg".}

\item{reg.col}{character, a meta-data column name which contains regulatory
region locations for the reg2 object.}
}
\value{
GRanges object with interacting regions from the first datasets 
that was present/confirmed by the second dataset of interacting regions. 
Meta-data from both input GRanges objects (storing info about interacting 
regions in the genome) are included in the final GRanges object, and 
meta-data column names are altered such that they indicate their origin (
eg from Reg1 object or Reg2 object).
}
\description{
A function that compares four input regions,eg two interacting regions
from one dataset with two interacting regions from the other dataset
in the criss-cross manner. THUS, the input orientation of the 
interacting regions is NOT strict: the first GRanges object of the region1
from one dataset will be compared to both; the first GRanges object and the
second GRanges object of region2(and its pair is compared in the opposite 
direction).
}
\details{
Function performs criss-cross overlap between 2 GRanges objects with
2 pairs of GRanges each. It performs reg2gene::OverlapRegions 
overlap between  (reg1:Coord1-reg2:Coord1)&(reg1:Coord2-reg2:Coord2) and 
vice-versa (reg1:Coord1-reg2:Coord2)&(reg1:Coord2-reg2:Coord1).
Switch is achieved by swaping coordinated of reg2 (eg positions of gene and
regulatory region locations stored as GRanges object within GRanges object 
are switched).
}
\author{
Inga Patarcic
}
\keyword{internal}
