% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkScriptsGRAdjusted.R
\name{confusionMatrix}
\alias{confusionMatrix}
\title{ConfusionTable statistics for the benchmarked reg2Gene object}
\usage{
confusionMatrix(reg2GeneBench, benchCol = "BenchmarkO", thresholdID = NULL,
  thresholdValue = 0.05, statistics = "PPV")
}
\arguments{
\item{reg2GeneBench}{Granges object which contains a metacolumn(type:logical) 
with the results of benchmarking procedure. It can be a benchmarked 
reg2Gene object resulted from \code{benchmark}.
!WARNING! Prior this analysis, it is advised to reduce the number of 
true negatives by including only reg2gene entries that could be potentially
benchmarked. THUS run \code{Filter_PreBench} prior running \code{benchmark})}

\item{benchCol}{character (default "benchmarkO"), a column name of metadata
that indicates the column where the result of benchmarking procedure is 
stored (as vector of TRUE and FALSE entries).}

\item{thresholdID}{character (def:NULL) name which indicates a column where
statistics of the modelling procedure is stored. This column is filtered such
that everything below predefined threshold is considered to be statistically 
significant association (set to be equal to 0), whereas everything above that
threshold is 0".}

\item{thresholdValue}{numeric (def:0.05) A value of a threshold.}

\item{statistics}{character (def "ConfusionMatrix"). Currentlty, 
"ConfusionMatrix" or "PPV" are options.}
}
\value{
Integer or list. If "ConfusionMatrix" is requested then the output is 
list with following values:TP,FP,TN,FN,Specificity, Accuracy, PPV, NPV,F1.
Otherwise only PPV is reported.
}
\description{
A function takes as an benchmarked reg2gene object and outputs
confusion table statistics.
}
\details{
Reports statistics based on the confusion matrix. 
There are four different categories in the confusion matrix: 
TP = (number of) true positive: reg2gene entry that was reported to be 
associated (reported gene-enhancer statistics lower than a predefined
threshold) and was benchmarked. 
FP = (number of) false positive: reg2gene entry that was reported to be
associated (reported gene-enhancer statistics lower than a predefined 
threshold) BUT was not overlapped with benchmark dataset
FN = (number of) false negative: reg2gene entry that was NOT reported to
be associated (reported gene-enhancer statistics NOT lower than a predefined 
threshold) BUT is benchmarked
TN = (number of) true negative: reg2gene entry 
that was NOT reported to be associated (reported gene-enhancer
statistics NOT lower than a predefined threshold) AND is NOT benchmarked.

The formulas used in this function are: \deqn{Sensitivity = TP/(TP+FN)} 
\deqn{Specificity = TN/(TN+FP)} 
\deqn{Accuracy = (TP+TN)/(TP+FN+FP+TN)}
\deqn{PPV = TP/(TP+FP)} 
\deqn{NPV = TN/(TN+FN)} 
\deqn{F1 = (2*TP)/((2*TP)+FP+FN)}
}
\examples{
confusionMatrix(BenchMarkedReg2Gene_toy,thresholdID = "PValue",
thresholdValue = 0.05,benchCol = "BenchmarkO",statistics = "ConfusionMatrix")
confusionMatrix(BenchMarkedReg2Gene_toy,thresholdID = "PValue",
  thresholdValue = 0.5,benchCol = "BenchmarkO",statistics = "PPV")

     
}
