% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkScriptsGRAdjusted.R
\name{ConfusionMatrixReg2Gene}
\alias{ConfusionMatrixReg2Gene}
\title{ConfusionTable statistics of Benchmarked Reg2Gene Object}
\usage{
ConfusionMatrixReg2Gene(BenchMarkedReg2Gene, thresholdID = NULL,
  thresholdValue = 0.05, statistics = "PPV")
}
\arguments{
\item{BenchMarkedReg2Gene}{A result of \code{BenchMarkReg2Gene}.
!WARNING! It is advised to reduce the number of true negatives by including only
Reg2gene entries that could be potentially benchmarked. THUS run \code{Filter_PreBench}
prior running \code{BenchMarkReg2Gene})}

\item{thresholdID}{character (def:NULL) name which indicates a column which 
will be filtered}

\item{thresholdValue}{numeric (def:0.05) A value of a threshold. Everything below 
that thresold will be consider as a positive, an is set to be equal to 1.
Otherwise 0.}

\item{statistics}{character (def "ConfusionMatrix"). Currentlty, "ConfusionMatrix" or
"PPV" are options. If "ConfusionMatrix" is requested then output of 
\code{\link[caret]{confusionMatrix}} is reported. Otherwise, if "PPV" 
is requested then output of \code{\link[caret]{posPredValue}}}
}
\value{
Reports either positive predictive value or statictics based on 
confusion matrix \code{\link[caret]{confusionMatrix}}
}
\description{
A function takes as an input results of benchmarking procedure for 
Reg2Gene objects \code{BenchMarkReg2Gene} and outputs
confusion table statistics.
}
\details{
Reports statistics based on the confusion matrix. There are four different categories
in the confusion matrix: 
1) TP (true positive): reg2gene entry that was reported to be associated (reported gene-enhancer
statistics lower than a predefined threshold) and is benchmarked.
2) FP (false positive): reg2gene entry that was reported to be associated (reported gene-enhancer
statistics lower than a predefined threshold) BUT does not overlap benchmark dataset
3) TN (true negative): reg2gene entry that was NOT reported to be associated (reported gene-enhancer
statistics NOT lower than a predefined threshold) AND is NOT benchmarked.
4) FN (false negative): reg2gene entry that was NOT reported to be associated (reported gene-enhancer
statistics NOT lower than a predefined threshold) BUT is benchmarked.

# @examples ConfusionMatrixReg2Gene(BenchMarkedReg2Gene=BenchmarkedReg2Gene_toy,
# thresholdID = "PValues",thresholdValue = 0.05)

#ConfusionMatrixReg2Gene(BenchmarkedReg2Gene_toy,thresholdID = "PValues",
# thresholdValue = 0.3, statistics = "ConfusionMatrix")
}
