% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associateReg2Gene.R
\name{corMat}
\alias{corMat}
\title{Correlation between a col matrix and a vector}
\usage{
corMat(y, mat, method = "pearson")
}
\description{
A convenience function that computes the the correlation of a given vector
with all the other column vectors of a matrix
}
\keyword{internal}
