% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associateReg2Gene.R
\name{corResample}
\alias{corResample}
\title{Correlation with resampling p-values}
\usage{
corResample(mat, method = "pearson", col = 1, B = 1000)
}
\arguments{
\item{mat}{matrix of response and predictor variables}

\item{method}{pearson or spearman}

\item{col}{column number of response variable}

\item{B}{number of random samples (shuffling) of given column}
}
\description{
private function that calculate correlation of given
column matrix with a column and assigns significance to
correlation coefficents based on resampling
}
\examples{

m=readRDS("data/sample_rawActivityMatrices/Roadmap/H3K27ac/ENSG00000140718_FTO.rds")

m=apply(m, 2,as.numeric) # change to numeric matrix
corResample(m,method="pearson",col=1,B=1000)

}
\keyword{internal}
