% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg2gene_plotF.R
\name{createIntTrack}
\alias{createIntTrack}
\title{Help function to create InteractionTrack}
\usage{
createIntTrack(x, interactionDataName, chr, coloring)
}
\description{
Help function to create InteractionTrack
}
\author{
IngaPa
}
\keyword{internal}
