% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associateReg2Gene.R
\name{dcorMat}
\alias{dcorMat}
\title{Distance correlation between a col matrix and a vector}
\usage{
dcorMat(y, mat)
}
\description{
a convenience function that computes the the distance correlation of a given
vector (y) with all the other columns of a matrix (mat)
}
\keyword{internal}
