% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associateReg2Gene.R
\name{detectGeneEnhPresence}
\alias{detectGeneEnhPresence}
\title{GRanges contains gene+enhancer info}
\usage{
detectGeneEnhPresence(x)
}
\arguments{
\item{x}{GRanges object - which contains info about gene expression and
enhancer activity}
}
\description{
Function that tests each GRanges object in GRangesList that it contains 
necessary info to run associateReg2Gene(): info about gene expression and
enhancer activity
}
\author{
Inga Patarcic
}
\keyword{internal}
