% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associateReg2Gene.R
\name{estimateGammaPval}
\alias{estimateGammaPval}
\title{Estimate P-values from resampling statistics using Gamma distribution}
\usage{
estimateGammaPval(vals, orig, abs = TRUE, add = 0)
}
\arguments{
\item{vals}{a vector or matrix of column vectors. Each column corresponds
to coefficients obtained from resampling fit}

\item{orig}{a single value or a vector of original coefficients obtained from the
original matrix}

\item{abs}{if TRUE, use absolute values of coefficients to fit a distribution}

\item{add}{an numeric value defaults to 0. if more than 0, add that value to
orig and vals. This way gamma distribution can model null distributions that
have negative values.}
}
\description{
The function estimates gamma distribution based P-values from the null
distribution obtained from resampling, a gamma distribution is fit
to the null distribution and p-values are calculated based on that fitted
distribution
}
\keyword{internal}
