% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataIntegRFunctions.R
\name{exonExpressionStrandAdjusted}
\alias{exonExpressionStrandAdjusted}
\title{Function that quantifies exon expression based on strand specificity}
\usage{
exonExpressionStrandAdjusted(exonsForQuant, exonsSplitted, target, sampleIDs,
  libStrand, mc.cores, summaryOperation)
}
\arguments{
\item{target}{character, paths to bigwig files}

\item{sampleIDs}{character, name of samples/bigwig files}

\item{libStrand}{character, library strandness ("+","-","*")}

\item{mc.cores}{integer, number of cores for mclapply}

\item{summaryOperation}{argument for ScoreMatrixBin() how to quantify scores}

\item{exons}{GRanges object with exon coordinates and metadata about genes}

\item{exons.splitted}{GRanges object with exon coordinates splitted based on
the strand.}
}
\value{
Granges object with metadata which corresponds to quantified scores 
per input exon regions and bigwig files
}
\description{
Function that separates stranded and unstranded libraries and quantifies 
 exon expression separately for stranded libraries eg expression of exons 
 located on the + strand is quantified using forward libraries if libraries
 are unstranded then exon orientations is not taken into account
}
\author{
Inga Patarcic
}
\keyword{internal}
