% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstreamFunctions.R
\name{extractInfo}
\alias{extractInfo}
\title{Help f() to extract info about cohort/method/algorithm}
\usage{
extractInfo(path, typeOptions = method)
}
\description{
Help f() to extract info about cohort/method/algorithm
}
\author{
Inga Patarcic
}
\keyword{internal}
