% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilteringFunctions.R
\name{FilterPerGeneModel}
\alias{FilterPerGeneModel}
\title{Filters genes and enhancer regions prior modelling}
\usage{
FilterPerGeneModel(TSSmodel)
}
\arguments{
\item{TSSmodel}{a GRanges object, member of the GRangesList from \code{regActivityAroundTSS}}
}
\value{
a vector of TRUE,FALSE entries which indicates whether ranges in the input
GRanges object passed filtering or not.
}
\description{
The function that takes as an input individual GRanges object stored in GRangesList from
\code{regActivityAroundTSS} and returns a vector of TRUE,FALSE entries corresponding to
members of GRange object and whether they pass filtering or not.
}
\details{
It checks whether genes and enhancer region have:a) an expression 
or enhancer activity eqaul to 0 across all samples(cell types) or 90% of expression values or
enhancer activity is equal to 0; b) variance across samples is different from 0.
FALSE value indicates that these ranges should filtered prior the modelling procedure; 
TRUE indicates that it is ok to proceed with this ranges. 
!IMPORTANT NOTE! If the first entry is FALSE, than that gene should be removed from the analysis
}
