% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkGInteractions.R
\name{filterPreBenchGI}
\alias{filterPreBenchGI}
\title{Preadjustiong for high number of true negatives in regulatoryReg-TSS pairs}
\usage{
filterPreBenchGI(reg2Gene, benchData, nCores = 1, ignore.strand = FALSE,
  ...)
}
\arguments{
\item{reg2Gene}{a \code{\link[InteractionSet]{GInteractions}} object output
from \code{\link{associateReg2Gene}}). Usually, 1st GRanges object,or anchor1
corresponds to the enahncer location, whereas the other GRanges object 
corresponds to the regulatory region locations.}

\item{benchData}{a \code{\link[InteractionSet]{GInteractions}} object output
from \code{\link{associateReg2Gene}}) or a list of GInteractions object. 
Both regions are used in the benchmarking procedure. 
This object stores benchmarking informations eg interacting region
coordinates from techniques such as HiC,eQTL studies...}

\item{nCores}{possible to be runned in parallel. Argument for mclapply f();
how many cores to use.}

\item{ignore.strand}{argument to be passed to
\code{\link[IRanges]{findOverlaps}}. When set to TRUE, the strand 
information is ignored in the overlap analysis.}

\item{...}{further arguments to methods, not implemented yet}
}
\value{
GInteractions object with added benchmark results metadata [Filter 
column].Each column metadata column corresponds to one benchmark dataset 
analyzed if input is list() 
Values are either 0/1 (possible or not to benchmark).
}
\description{
By identifying entries that could be potentially benchmarked this 
function eliminates all regulatory region-TSS pairs [anchor1 and anchor2 from
reg2Gene] that do not overlap with any benchmark anchor1 or anchor2 location. 
E.g. it selects reg2Gene regions only when both regulatory region and TSS  
have overlapping regions somewhere in the benchmarking set; across all 
benchmark anchor pairs, and not necessarily overlapping regions of the same
benchmark pair.
This is useful to improve the confusion matrix statistics reported by 
\code{confusionMatrix} by eliminating the high number of true negatives. 
TN are very abundant in the reg2gene dataset since benchmark dataset usually
covers much smaller regions of the genome (method limitations)
}
\details{
All benchmark regions that can be confirmed by any combination of 
selected unique union of anchor1 or anchor2 form reg2Gene object is used
as anchor1-anchor2 pairs that can be benchmarked. Reasoning, if present in
this set anchor1 or anchor2 regions form reg2Gene object necessarily need to
have other member of the pair overlapping somewhere in benchmark dataset.
}
\examples{
# Creating testing and benchmarking dataset

require(GenomicRanges)
require(InteractionSet)
   
   reg2Gene <- GInteractions(GRReg1_toy,GRReg1_toy$reg)
   benchData <- GInteractions(GRReg2_toy,GRReg2_toy$reg)
   
 filterPreBenchGI(reg2Gene,
                  benchData,
                  nCores=1)
                  
##################   
# example for list:

# NOTE: anchor1.Bench1Exp & anchor1.Bench2Exp are expected/precalculated
# values for this benchmark example
 
  benchDataList <- list(benchData,reg2Gene)
  names(benchDataList) <- c("fData1","fData2")
  reg2GeneBF.list <-filterPreBenchGI(reg2Gene,
                                     benchDataList,
                                     nCores = 1)
 
 ####################################
 
 # Checking what happends when anchor1&anchor2 both overlap only one region
 # in benchmark dataset?  OK, They are filtered OUT...
 
   filterPreBenchGI(reg2Gene[1],
                    benchData[1],
                    nCores = 1) 
    
  
}
