% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkGInteractions.R
\name{filterPreBenchGIsimple}
\alias{filterPreBenchGIsimple}
\title{Filtering help function}
\usage{
filterPreBenchGIsimple(interactions, benchInteractions, ignore.strand = FALSE)
}
\description{
Filter interactions models using benchmark data but for only one GInteraction
object [not lists]
}
\details{
Read description for \code{filterPreBenchGI}
}
\author{
Inga Patarcic
}
\keyword{internal}
