% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associateReg2Gene.R
\name{getGeneEnhScoresDF}
\alias{getGeneEnhScoresDF}
\title{extract gene-enhancer scores from GRanges object}
\usage{
getGeneEnhScoresDF(gr)
}
\arguments{
\item{gr}{perGene GRanges object}
}
\description{
Extract gene-enhancer scores from metadata of GRanges object and filters NA
values using stepwise.complete.cases()
}
\author{
Inga Patarcic
}
\keyword{internal}
