% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotF.R
\name{getStatistics}
\alias{getStatistics}
\title{Selects column that will be used as an input to plot heigth of the loops,
eg "pval". However -log10(pvalue) is calculated, not original value}
\usage{
getStatistics(x, statistics)
}
\description{
Selects column that will be used as an input to plot heigth of the loops,
eg "pval". However -log10(pvalue) is calculated, not original value
}
\author{
IngaPa
}
\keyword{internal}
