% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotF.R
\name{gettingGeneInfoEnsembl}
\alias{gettingGeneInfoEnsembl}
\title{Help function to get info about exon location for genes of interest from
Ensembl}
\usage{
gettingGeneInfoEnsembl(GeneList, filters)
}
\description{
filters <- c("hgnc_symbol","ensembl_gene_id")
}
\author{
IngaPa
}
\keyword{internal}
