% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associateReg2Gene.R
\name{glmnetResample}
\alias{glmnetResample}
\title{internal function that uses glmnet for prediction}
\usage{
glmnetResample(mat, col = 1, B = 1000, ...)
}
\arguments{
\item{mat}{matrix of response and predictor variables}

\item{col}{column number of response variable}

\item{B}{number of random samples (shuffling) of response variables}

\item{...}{further arguments to glmnet, not implemented yet}
}
\description{
private function that uses glmnet to predict fiven column from the rest
and assigns significance to coefficents based on resampling
}
\examples{

m=readRDS("data/sample_rawActivityMatrices/Roadmap/H3K27ac/ENSG00000140718_FTO.rds")

m=apply(m, 2,as.numeric) # change to numeric matrix
glmnetResample(scale(m),col=1,B=1000)
}
\keyword{internal}
