% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associateReg2Gene.R
\name{grlist2GI}
\alias{grlist2GI}
\title{convert GRangesList to GInteractions}
\usage{
grlist2GI(grlist)
}
\arguments{
\item{grlist}{a GRangesList output from \code{\link{regActivityAroundTSS}}}
}
\description{
The function converts output of \code{\link{regActivityAroundTSS}} (a
GRangesList) to a GInteractions object by rearranging the rows and columns so
that output GInteractions object contains regulatory region coordinates 
and an additional column "reg" which contains the gene info
}
\author{
Altuna Akalin, Inga Patarcic
}
\keyword{internal}
