% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associateReg2Gene.R
\name{grlist2gr}
\alias{grlist2gr}
\title{convert GRangesList to GRanges with associated regulatory regions (internal)}
\usage{
grlist2gr(grlist)
}
\arguments{
\item{grlist}{a GRangesList output from \code{\link{regActivityAroundTSS}}}
}
\description{
The function converts output of \code{\link{regActivityAroundTSS}} (a
GRangesList) to a GRanges object by rearranging the rows and columns so
that output GRanges object mainly contains regulatory region coordinates 
and an additional column "reg" which contains the gene info
}
\author{
Altuna Akalin, Inga Patarcic
}
\keyword{internal}
