% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GRReg1_toy}
\alias{GRReg1_toy}
\title{Sample file for Region1 imported as GRanges}
\format{GRanges object with 6 ranges and 3 metadata columns}
\usage{
GRReg1_toy
}
\description{
Sample file for Region1 imported as GRanges
}
\keyword{datasets}
