% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GRReg2_toy}
\alias{GRReg2_toy}
\title{Sample file for benchmark dataset}
\format{GRanges object with 14 ranges and 3 metadata columns}
\usage{
GRReg2_toy
}
\description{
Sample file for benchmark dataset
}
\keyword{datasets}
