% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstreamFunctions.R
\name{makeIndexTable}
\alias{makeIndexTable}
\title{Creates Index table of files produced by reg2gene data
integration and modelling}
\usage{
makeIndexTable(pathModels, type = "ind", method = "H3K4me1",
  algorithm = "pearson", cohort = "Roadmap")
}
\arguments{
\item{pathModels}{(character) a path to folder where .rds files are stored or 
an object with paths to .rds files.  
# IMPORTANT! 
cohort/method/algorithm names should be indicated in the path, 
or .rds files with this name should be present in the defined directory}

\item{type}{"ind" (default; character)
OPTIONS: ("votingAlgoritm","votingCohorts","votingMethods","metaA","regAct",
"ind").
Define whether results are produced by voting by("votingAlgoritm",
"votingCohorts","votingMethods"), meta-analysis("metaA"), or from individual 
modelling ("ind") or a result of regActivityAroundTSS() ("regAct"). 
Important for analyses when all these info is gathered together}

\item{method}{"H3K4me1" (default; character). A character vector of all 
methods used for this analysis. 
Written for the option: method=c("H3K4me1","H3K27ac","Methylation","DNase")}

\item{algorithm}{"pearson" (default; character). A character vector of all 
methods used for this analysis. 
Written for the option: algorithm=c("pearson","spearman","elasticnet","dcor",
"randomForest")}

\item{cohort}{"Roadmap" (default; character). A character vector of all 
cohorts used for this analysis. 
Written for the option: cohort= c("Roadmap","Blueprint","CEMT","McGill")}

\item{...}{maybe to add in the future}
}
\value{
A data frame with the following columns (n=5):
1) path to .rds files - results produced by modelling or data integration 
re2gene f() and corresponding info:
2) methods - which method type is used for this analysis: eg. H3K4me1,...
3) algorithms - which algorithm type is used for this analysis: eg. dcor,...
4) cohorts - which cohort type is used for this analysis: eg. Roadmap,...
5) type - define whether results are produced by voting("votingAlgoritm",
"votingCohorts","votingMethods"), meta-analysis("metaA"),  from individual 
modelling ("ind") or a result of regActivityAroundTSS() ("regAct")
}
\description{
The function creates a data frame with the following columns:
1) path to .rds files - results produced by modelling or data integration 
re2gene f() and corresponding info:
2) methods - which method type is used for this analysis: eg. H3K4me1,...
3) algorithms - which algorithm type is used for this analysis: eg. dcor,...
4) cohorts - which cohort type is used for this analysis: eg. Roadmap,...
5) type - define whether results are produced by voting, meta-analysis,from 
individual modelling  or a result of regActivityAroundTSS()
}
\details{
An IndexTable which is a result of makeIndexTable() is used as
an input for following functions: selectEP(),votedEP(),plotGEEA() to
gathers paths to different reg2gene data integration and modelling results,
and allows easier downstream analysis. This f() should be separately runned
for voting("votingAlgoritm","votingCohorts","votingMethods"), 
meta-analysis and individual modelling, and regActivityAroundTSS() results, 
and individual IndexTables should be pooled together, such that plotGEEA()
function can be runned (it requires paths to "regAct" .rds files - results of 
regActivityAroundTSS() and for example modelling paths.
}
\examples{
require(stringr)

makeIndexTable(pathModels="~/H3K4me1pearsonRoadmap.rds",
type="ind",method="H3K4me1",algorithm="pearson",cohort="Roadmap")

# cohort/method/algorithm names should be indicated in the path, or .rds 
# files with this name should be present in the defined directory

makeIndexTable(pathModels="~/test.rds",
type="ind",method="H3K4me1",algorithm="pearson",cohort="Roadmap")


# HOW-TO-USE-IT EXAMPLE

IndexTable1 <- makeIndexTable(pathModels="~/H3K4me1pearsonRoadmap.rds",
type="ind",method="H3K4me1",algorithm="pearson",cohort="Roadmap")

IndexTable2 <- makeIndexTable(pathModels="~/ra_H3K4me1pearsonRoadmap.rds",
type="regAct",method="H3K4me1",algorithm="pearson",cohort="Roadmap")

IndexTable <- rbind(IndexTable1,IndexTable2) # IndexTable use downstream  

\dontrun{
algorithm <- c("pearson","spearman","elasticnet","dcor","randomForest")
cohort=c("Roadmap","Blueprint","CEMT","McGill")
method=c("H3K4me1","H3K27ac","Methylation","DNase")

votedPath="/data/akalin/Projects/AAkalin_Catalog_RI/Results/Validation/Fishillevic/VoteD/"
metaPath="/data/akalin/Projects/AAkalin_Catalog_RI/Results/Validation/Fishillevic/MetaA/"
pathRegActTSS="/data/akalin/Projects/AAkalin_reg2gene/Results/regActivityAroundTSS/FishilevichChromHMMoneTOone/"
pathModels="/data/akalin/Projects/AAkalin_Catalog_RI/Data/ValidationDataset/Fishillevich/ModellingResults/"
votedAlg="/data/akalin/Projects/AAkalin_Catalog_RI/Results/Validation/Fishillevic/VoteD/VotingAlg/"
votedC="/data/akalin/Projects/AAkalin_Catalog_RI/Results/Validation/Fishillevic/VoteD/VotingCohort/"
votedM="/data/akalin/Projects/AAkalin_Catalog_RI/Results/Validation/Fishillevic/VoteD/VotingMethod/"

IndexTable <- data.frame(rbind(makeIndexTable(metaPath,"metaA"),
                               makeIndexTable(votedAlg,"votingAlgorithm"),
                               makeIndexTable(votedC,"votingCohorts"),
                               makeIndexTable(votedM,"votingMethods"),
                               makeIndexTable(pathRegActTSS,"regAct"),
                               makeIndexTable(pathModels,"ind")),stringsAsFactors = F)
saveRDS(IndexTable,"/data/akalin/Projects/AAkalin_Catalog_RI/Data/ValidationDataset/Fishillevich/IndexTable.rds")

}

}
\author{
Inga Patarcic
}
\keyword{internal}
