% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associateReg2Gene.R
\name{manyZeros}
\alias{manyZeros}
\title{decide if the input colum from a matrix has many zeros}
\usage{
manyZeros(mat, col = 1)
}
\arguments{
\item{mat}{a column matrix that contains gene expression values and
regulatory region activities.}

\item{col}{column number of response variables}
}
\description{
The input column matrix represent gene expression and regulatory region
activities. decides if the genes
}
\examples{

m=readRDS("data/sample_rawActivityMatrices/Roadmap/H3K27ac/ENSG00000140718_FTO.rds")

}
\author{
Altuna Akalin
}
\keyword{internal}
