% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaAssociations.R
\name{metaAssociations}
\alias{metaAssociations}
\title{meta-analysis for regulatory region and gene associations}
\usage{
metaAssociations(associations)
}
\arguments{
\item{associations}{A GRangesList where each element is a GRanges
object output from \code{\link{associateReg2Gene}} function.}
}
\value{
returns a \code{\link[GenomicRanges]{GRanges}} object with
updated statistics
from the meta-analysis. The output will be similar to
the output of \code{\link{associateReg2Gene}} function, it
will have the same meta-data columns in the same order.
}
\description{
The function combines association P-values and coefficients from
different data
sets using Fisher's method and weigthed averaging respectively. After,
combining P-values, q-values are calculated using the
\code{\link[qvalue]{qvalue}} function.
}
\examples{
\dontrun{
#
assoc=system.file("extdata", "sampleAssociationsGrlist.rds",
                      package = "reg2gene")

meta=metaAssociations(assoc)
}

}
\author{
Altunislav Akalinski
}
