% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaAssociationsGI.R
\name{metaInteractions}
\alias{metaInteractions}
\title{meta-analysis for regulatory region and gene interactions}
\usage{
metaInteractions(interactions)
}
\arguments{
\item{interactions}{A list of \code{\link[InteractionSet]{GInteractions}}
objects outputed from \code{\link{associateReg2Gene}} function.}
}
\value{
returns a \code{\link[InteractionSet]{GInteractions}} object with an 
updated statistics from the meta-analysis. The output will be similar to
the output of \code{\link{associateReg2Gene}} function, it
will have the same meta-data columns in the same order.
}
\description{
The function combines association P-values and coefficients from different 
data sets using Fisher's method and weigthed averaging respectively. 
It is useful to combine datasets produced by different research groups while
avoiding problems such as batch effects. 
For example, it can be used to combine Roadmap and Blueprint 
\code{\link{associateReg2Gene}} gene~enhancer single-model information 
obtained using across cell RNA-Seq signals and CHiP-Seq H3K27ac tracks.
Aggregating single-model information (gene-enhancer pairs) by means of 
meta-analysis across different data-sources should increase the statistical 
power, improve the precision and accuracy of estimates and altogether 
produce more robust and reproducible results. After,combining P-values, 
q-values are calculated using the \code{\link[qvalue]{qvalue}} function.
}
\examples{
# creating datasets

require(GenomicRanges)
require(InteractionSet)

# INPUT 0: 2 GRanges objects with toy expression values (x,y,z,a)

gr2 <- gr <- GRanges(seqnames=rep("chr1",3),IRanges(1:3,3:5))
   x <- 1:5
   y <- 2:6
   z <- 10:14
   a <- rep(0,length(x))
   
   
   GeneInfo <- as.data.frame(matrix(c(rep("gene",3),rep("regulatory",6)),
               ncol = 3,byrow = TRUE),stringsAsFactors=FALSE)
               colnames(GeneInfo) <- c("featureType","name","name2")
 
 mcols(gr) <- DataFrame(cbind(GeneInfo,rbind(x,y,z)))
 mcols(gr2) <- DataFrame(cbind(GeneInfo,rbind(x,y,a)))
 
# RUNNING associateReg2Gene and obtaining output results
 
 AssocObject <- reg2gene::associateReg2Gene(gr)
 AssocObject2 <- reg2gene::associateReg2Gene(gr2)
 
 # input for meta-analysis is list of such objects
 
 interactions <- list(AssocObject,AssocObject2)
 names(interactions) <- c("AssocObject","AssocObject2")
 
 # OUTPUT: Run metaA
 
 metaInteractions(interactions)

}
\author{
Altunislav Akalinski
}
