% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstreamFunctions.R
\name{metaVotePoStat}
\alias{metaVotePoStat}
\title{Reports a number of statistically significant enhancer~
promoter associations}
\usage{
metaVotePoStat(indexTable = IndexTable, type = "votingAlgorithm",
  typeSuccess = "pval", thresholdSuccess = 0.1, method = "H3K4me1",
  algorithm = "dcor", cohort = "Roadmap")
}
\arguments{
\item{indexTable}{an output of the makeIndexTable() for given combination of
methods/algorithms/cohorts; a table should contain the following columns: 
path - paths to .rds files; methods - a method type used for the analysis: 
eg. H3K4me1; 
algorithms -  algorithm used for the analysis: eg. dcor, 
cohorts - cohort used for the analysis: eg. Roadmap; 
type - define whether results are produced by voting or meta-analysis, 
AND individual modelling (indexTable$type=="ind"). 
It is necessary that individual modelling paths are
present in the index table (indexTable$type=="ind"), because results of these
models are compared to the results of voting or meta-analysis procedure.}

\item{type}{"votingAlgorithm" (default; character). Describes which voting
procedure is used to produce voting results (common options:"votingAlgoritm",
"votingCohorts","votingMethods") or if meta-analysis is used (type="metaA").
This argument corresponds to the type argument in 
\code{\link{makeIndexTable}}, thus if some other type character object is
used with \code{\link{makeIndexTable}} then the same argument should be used}

\item{typeSuccess}{"pval" (default) Other options "qval". Which statistics to
threshold to assess a statistical significance.}

\item{thresholdSuccess}{0.05 (default, numeric). A threshold useD to 
assess a statistical significance.}

\item{method}{"H3K4me1" (default; character). For which method to 
extract Enh~Promoter pairs. Common options: "H3K4me1","H3K27ac",
"Methylation","DNase"}

\item{algorithm}{"pearson" (default; character). For which algorithm to
extract Enh~Promoter pairs. Common options:"pearson","spearman",
"elasticnet","dcor","randomForest"}

\item{cohort}{"Roadmap" (default; character). For which cohort  to extract
Enh~Promoter pairs. Common options:"Roadmap","Blueprint",
"CEMT","McGill"}
}
\value{
returns a dataframe with information about the number of identified 
statistically enhancer~promoter associations and info about corresponding
input modelling datasets used - cohort,method,algorithm combination.
The same statistics is returned for the requested voting method or 
meta-analysis.
}
\description{
metaVotePoStat() reports the number of statistically significant enhancer~
promoter associations for voting or meta-analysis and corresponing individual
modelling results. IMORTANT NOTE: since it uses results of previously runned
voting procedure it is important that statistics and threshold values used in
voting procedure are the same as used in this algorithm. However,
meta-analysis statistics thresholding is performed as a part of this function
,thus statistics and threshold [typeSuccess&thresholdSuccess] values are used
to filter both: meta-analysis results and individual modelling results.
}
\details{
This function allows an easy assesment of the success of voting or 
meta-analysis procedure compared to individual modelling procedures, through 
returning the number of statically significant cases reported in voting or 
meta-analysis and corresponding individual modelling procedures.
}
\examples{
library(stringr)
library(InteractionSet)
\dontrun{

IndexTable <- readRDS("/data/akalin/Projects/AAkalin_Catalog_RI/Results/Validation/Fishillevic/VoteD//CohortVoting_McGillH3K4me1.rds")

metaVotePoStat(indexTable=IndexTable, 
type="votingAlgorithm",
method="H3K4me1", 
algorithm="dcor",
cohort="Blueprint")

}

}
\author{
Inga Patarcic
}
