% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ModellingTest}
\alias{ModellingTest}
\title{GRanges example to test modelling of associateReg2gene function from reg2gene}
\format{GRanges object with 9 ranges and 55 metadata columns:}
\usage{
ModellingTest
}
\description{
GRanges example to test modelling of associateReg2gene function from reg2gene
}
\details{
cor(y~x1)=0.99; cor(y~x2)=0.9,cor(y~x3)=0.8,...,cor(y~x8)=0
Drop in correlation can be used to test results of modelling procedure. 
52 "celltypes" included to approximate for 52 H3K4me1&RNASeq datasets from 
Roadmap
}
\keyword{datasets}
