% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkScriptsGRAdjusted.R
\name{overlapRegions}
\alias{overlapRegions}
\title{Overlap of interacting regions}
\usage{
overlapRegions(reg1, reg2)
}
\arguments{
\item{reg1}{- GRanges object with at least one meta-data column which stores
the second GRanges object. Usually,one GRanges object corresponds to the gene 
locations, preferably TSS location, whereas the other GRanges object 
corresponds to the regulatory region locations. Meta-data column name should
be "reg".}

\item{reg2}{- GRanges object with at least one meta-data column which stores
the second GRanges object.  Meta-data column name should be "reg".}
}
\value{
GRanges object with interacting regions from the first datasets 
that was present/confirmed by the second dataset of interacting regions. 
Meta-data from both input GRanges objects (storing info about interacting 
regions in the genome) are included in the final GRanges object, and 
meta-data column names are altered such that they indicate their origin (
eg from Reg1 object or Reg2 object).
}
\description{
The function that compares two sets of interacting regions in the genome 
and reportes interacting regions from the first datasets that is present/
confirmed by the second dataset of interacting regions.
}
\details{
Function that performs a simple findOverlapsoverlap twice - 
between 2 sets of coordinates stored.  It returns rows in which both 
combinations of overlaps are TRUE.
}
\author{
Inga Patarcic
}
\keyword{internal}
