% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associateReg2Gene.R
\name{perGeneModelling}
\alias{perGeneModelling}
\title{Function returns output of called modelling procedure indivudually for 
each GRanges object}
\usage{
perGeneModelling(x, mat, ModelRes)
}
\arguments{
\item{gr}{perGene GRanges object}
}
\description{
It extracts gene-enhancer scores from metadata of GRanges object and filters
NA values using stepwise.complete.cases() embedded in getGeneEnhScoresDF()
Performs FILTERING: drop gene and NA is returned due to low or zero variation
in gene expression or when no enhancers are remained after 
stepwise.complete.cases()
}
\author{
Inga Patarcic
}
\keyword{internal}
