% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstreamFunctions.R
\name{plotGEEA}
\alias{plotGEEA}
\title{Scatterplots of enhancer activity~gene expression for max 16 EP pairs}
\usage{
plotGEEA(enhPromPairs, indexTable, method = "H3K4me1", algorithm = "dcor",
  cohort = "Roadmap")
}
\arguments{
\item{enhPromPairs}{A GInteractions object of length 16 with the 
corresponding statistics as a results of either reg2gene modelling 
\code{\link{associateReg2Gene}}, meta-analysis \code{\link{metaAssociations}}
or their pre-filtered versions \code{\link{selectEP}}. 
It requires information about statistics 
("coef") and corresonding testing results ("pval"). In the future it could 
be adjusted for ("qval" or that it does not require any statistics).}

\item{indexTable}{an output of the makeIndexTable() for given combination of
methods/algorithms/cohorts; a table should contain the following columns: 
path - paths to .rds files; methods - a method type used for the analysis: 
eg. H3K4me1; 
algorithms -  algorithm used for the analysis: eg. dcor, cohorts - cohort
used for the analysis: eg. Roadmap; type - define whether results are 
produced by meta-analysis, from individual modelling or a result of
regActivityAroundTSS() [NECESSARY]}

\item{method}{"H3K4me1" (default; character). For which method to 
extract Enh~Promoter pairs. Common options: "H3K4me1","H3K27ac",
"Methylation","DNase"}

\item{algorithm}{"pearson" (default; character). For which algorithm to
extract Enh~Promoter pairs. Common options:"pearson","spearman",
"elasticnet","dcor","randomForest"}

\item{cohort}{"Roadmap" (default; character). For which cohort  to extract
Enh~Promoter pairs. Common options:"Roadmap","Blueprint",
"CEMT","McGill"}
}
\description{
For GInteractions object that contains associations between a regulatory 
regions and TSSs (EP pairs),estimated association statistic,and corresponding
p-values this function read files that contain quantified (and normalized) 
enhancer activity and gene expression levels for each data entry (cell type).
Then this info is plotted as a scatterplot of enhancer activity~gene 
expression for max 16 EP pairs.
}
\details{
This function allows quick visualization of selected EP pairs as a
scatterplot of corresponding levels of enhancer activity and gene expression 
for all data entries (cell types used in the modelling procedure:
\code{\link{associateReg2Gene}}). This info is plotted for max 16 EP pairs, 
thus only the first 16 pairs will be plotted.
It is useful in the case when one wants to visualize the background levels of 
enhancer activity and gene expression for TOP EP associations. Since, for 
each combination of cohort/methods/algorithms separated modelling is 
performed one needs to define a combination.
}
\examples{

require(ggplot2)
require(stringr)

\dontrun{

TOPgenes <- readRDS("/data/akalin/Projects/AAkalin_Catalog_RI/Results/Validation/Fishillevich/TOPgenesH3K4me1dcorRoadmap.rds")
IndexTable <- readRDS("/data/akalin/Projects/AAkalin_Catalog_RI/Results/Validation/Fishillevic/VoteD//CohortVoting_McGillH3K4me1.rds")

plotGEEA(TOPgenes,
       indexTable=IndexTable,
       method="H3K4me1",
       algorithm="dcor",
       cohort="Roadmap")

}
}
\author{
Inga Patarcic
}
\keyword{internal}
