% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataIntegRFunctions.R
\name{quantifyGeneExpression}
\alias{quantifyGeneExpression}
\title{Function that quantifies gene expression based on prequantified exon exp}
\usage{
quantifyGeneExpression(expressionPerGene, exonsForGene)
}
\arguments{
\item{expressionPerGene}{GRanges object with exon coordinates and 
metadata info about corresponding genes and quantifed exon expression}
}
\value{
Granges object of genes with metadata which corresponds to quantified
gene expression
}
\description{
Function that calculates per gene expression based on exon expression results
sum across exons (mean value across exon*exon width)/
gene_length(sum_of_all_exon_lengths)
}
\author{
Inga Patarcic
}
\keyword{internal}
