% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associateReg2Gene.R
\name{qvaluCal}
\alias{qvaluCal}
\title{internal qvalue calculation function}
\usage{
qvaluCal(comb.res)
}
\description{
private function calculates qvalue for 2 categories of pvalue
}
\keyword{internal}
