% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associateReg2Gene.R
\name{rfResample}
\alias{rfResample}
\title{internal RandomForest prediction function}
\usage{
rfResample(mat, scaleData = scaleData, col = 1, B = B)
}
\arguments{
\item{mat}{matrix of response and predictor variables}

\item{col}{column number of response variables}

\item{B}{number of random samples (shuffling) of response variables}
}
\description{
private function that uses randomForest to predict fiven column from the rest
and assigns significance to
correlation coefficents based on resampling
}
\keyword{internal}
