% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotF.R
\name{selectGeneF}
\alias{selectGeneF}
\title{Selects genes of interest and associated reg regions from GInteractions obj}
\usage{
selectGeneF(x, selectGene)
}
\description{
Selects genes of interest and associated reg regions from GInteractions obj
}
\author{
IngaPa
}
\keyword{internal}
