% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg2gene_plotF.R
\name{selectRegR}
\alias{selectRegR}
\title{Selects reg regions of interest from GInteractions obj,and associated genes
however for associated genes it reports as well all other interactions 
that are associated with genes associated with queried reg region}
\usage{
selectRegR(x, selectRegulatoryRegion)
}
\description{
Selects reg regions of interest from GInteractions obj,and associated genes
however for associated genes it reports as well all other interactions 
that are associated with genes associated with queried reg region
}
\author{
IngaPa
}
\keyword{internal}
