% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotF.R
\name{setColors}
\alias{setColors}
\title{Adjusts for colors of plot: three options:
1) all red, in not info provided
2) uses user-provided color vector (eg "color" column from GInteractions
object)
3) If user selects region of interest, this function still plots other 
interactions around, but only in grey, whereas identified interaction
supposed to be red}
\usage{
setColors(x, coloring)
}
\description{
Adjusts for colors of plot: three options:
1) all red, in not info provided
2) uses user-provided color vector (eg "color" column from GInteractions
object)
3) If user selects region of interest, this function still plots other 
interactions around, but only in grey, whereas identified interaction
supposed to be red
}
\author{
IngaPa
}
\keyword{internal}
