% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associateReg2Gene.R
\name{stepwise.complete.cases}
\alias{stepwise.complete.cases}
\title{Stepwise complete cases method}
\usage{
stepwise.complete.cases(x)
}
\arguments{
\item{mat}{a column matrix that contains gene expression values and
regulatory region activities.}
}
\description{
Removes NA but in e order: 1st cells that have NA for at least
3/4 of enhancer regions, then the same thing is applied for cell types
then individual NAs are removed
}
\author{
Inga Patarcic
}
\keyword{internal}
