% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkScriptsGRAdjusted.R
\name{switchReg}
\alias{switchReg}
\title{Function that swaps gene and regulatory region locations}
\usage{
switchReg(reg, regCol)
}
\arguments{
\item{reg}{GRanges object with gene locations and meta-data which stores
info about regulatory region locations (as a GRanges object)}

\item{reg.col}{character, a meta-data column name which contains regulatory
region locations}
}
\description{
Function that swaps 2 GRanges objects(one stored as a meta-data within the 
other).Thus,reg meta-data stores info about gene locations, wheres main 
ranges store ranges previously stored as a reg meta-data
}
\author{
Inga Patarcic
}
\keyword{internal}
