% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstreamFunctions.R
\name{votedEP}
\alias{votedEP}
\title{Returns voted EP pairs for the combination of cohort/method/
algorithm and the voting procedure.}
\usage{
votedEP(indexTable, votingType = "votingAlgoritm", method = "H3K4me1",
  algorithm = "pearson", cohort = "Roadmap")
}
\arguments{
\item{indexTable}{an output of the makeIndexTable() for given combination of
methods/algorithms/cohorts; a table should contain the following columns: 
path - paths to .rds files; methods - a method type used for the analysis: 
eg. H3K4me1; 
algorithms -  algorithm used for the analysis: eg. dcor, 
cohorts - cohort used for the analysis: eg. Roadmap; 
type - define whether results are produced by voting, meta-analysis, from
individual modelling or a result of regActivityAroundTSS()}

\item{votingType}{"votingAlgoritm" (default; character) Describes which
voting procedure was used to produce voting results.
OPTIONS: ("votingAlgoritm","votingCohorts","votingMethods").}

\item{method}{"H3K4me1" (default; character). For which method to 
extract Enh~Promoter pairs. Common options: "H3K4me1","H3K27ac",
"Methylation","DNase"}

\item{algorithm}{"pearson" (default; character). For which algorithm to
extract Enh~Promoter pairs. Common options:"pearson","spearman",
"elasticnet","dcor","randomForest"}

\item{cohort}{"Roadmap" (default; character). For which cohort  to extract
Enh~Promoter pairs. Common options:"Roadmap","Blueprint",
"CEMT","McGill"}
}
\description{
This function allows easy import of results of different voting procedures
for given combination of cohort/method/algorithm.
}
\details{
This function allows easy import of results of different voting
procedures for given combination of cohort/method/algorithm. For example, 
"votingAlgoritm" correspond to voting based on the results of pearson,
spearman,elasticnet,dcor,randomForest modelling procedures for a combination
of cohort and method (eg H3K4me1+Roadmap). In this case algorithm 
argument is ignored while importing "votingAlgoritm" results 
for H3K4me1+Roadmap
}
\examples{

\dontrun{
library(InteractionSet)
library(stringr)
IndexTable <- readRDS("/data/akalin/Projects/AAkalin_Catalog_RI/Results/Validation/Fishillevic/VoteD//CohortVoting_McGillH3K4me1.rds")

votedEP(indexTable,
votingType="votingAlgorithm",
method="H3K4me1",
algorithm="pearson",
cohort="Roadmap")
 
}
}
\author{
Inga Patarcic
}
