% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataIntegRFunctions.R
\name{bwToGeneExp}
\alias{bwToGeneExp}
\title{Quantifies gene expression measured by RNA-Seq}
\usage{
bwToGeneExp(Exons, GeneExpSignals, sampleIDs = NULL, LibStrand = NULL,
  summaryOperation = "mean", mc.cores = 1, normalize = "ratio")
}
\arguments{
\item{Exons}{A GRanges object that contains exon regions over which 
the expression will be calculated. A meta-columns with following info are
necessary: need to adjust these}

\item{GeneExpSignals}{a named list of RNA-Seq BigWig files. Names correspond to 
the unique sample ids/names. Stranded and unstranded libraries allowed.
 BUT!!! It is crucial that forward and reverse RNA-Seq libraries are listed
in a row (eg one on top of each other)}

\item{sampleIDs}{NULL (default). A vector of unique sample
ids/names(.bw files), ordered as the bigwig files are ordered. When NULL
basenames of .bw files is used as a unique sample ids/names.}

\item{LibStrand}{a vector of "*","+,"-" (default NULL) which needs to be entered
as an argument.This vector provides info about the order of RNA-Seq libraries
based on their strandness: "+" corresponds to forward/positive RNA-Seq bigwig files; 
"-" corresponds to reverse/negative RNA-Seq bigwig files and "*" is unstranded
library. When all libraries are unstranded then the vector should contain a list
of "*" with the lenght equal to the number of analyzed RNA-Seq libraries
(eg bigwig files). If LibStrand=NULL than function will do it automatically, eg
create a vector of "*".It is crucial that stranded RNA-Seq libraries are
listed in a row (eg one on top of each other)}

\item{summaryOperation}{"mean"(default).
This designates which summary operation should be used over the regions.
Currently, only mean is available, but "median" or "sum" will be implemented
in the future.}

\item{mc.cores}{(def:1) Define the number of cores to use;
at most how many child processes will be run simultaneously 
using mclapply from parallel package. Parallelization requires at 
least two cores.}

\item{normalize}{NULL(default). Optional "quantile" and "ration"
If set to "quantile" activity measures are quantile normalized as 
implemented in \code{\link[preprocessCore]{normalize.quantiles}} and
returned ; if set to "ration" then "median ration method" implemented
as \code{\link[DESeq2]{estimateSizeFactorsForMatrix}} is used to normalize 
activity measures.}
}
\value{
a GRanges object where its meta-columns correspond to quantified
gene expression across cell type or condition. GRanges correspond
to the TSS location of the analyzed gene. Additionally, each TSS 
contains following metadata:a "name" and "name2" columns for 
unique id or name/symbol for the gene which the TSS is associated with. 
One is Ensembl id and the other could be used for the gene symbol. 
Other metadata column names should represent sample names/ids and should
match the GRanges object provided via regActivity argument.
}
\description{
The function quantifies gene expression as a sum of exon 
expressions quantified over pre-defined exon regions using 
signal from RNA-Seq tracks (bigwig files) and returns a
GRanges object with TSS location and corresponding gene
expression levels quantified over a set of samples.
}
\details{
For a gene of interest and across all samples (.bw files)
individaully,the expression is firstly calculated for all exon 
regions which correspond to the gene of interest. Then, per gene
exon expression scores are summed together 
and divided by a total exon length. Normalization is runned
on the level of gene expression. Currently, the relevant bigWig 
files are required to calculate activity. This function might 
be extended to work with BAM files in the future. 
RNA-Seq .bw files can come from stranded,unstranded or mixed 
libraries.



#@examples load(file="~/GeneExpSignals.RData")
#load(file="~/LibStrand.RData")
#load(file="~/Exons.RData")
#load(file="~/sampleIDs.RData")

 #   bwToGeneExp(Exons,GeneExpSignals,LibStrand,
 #    sampleIDs=NULL,mc.cores=1,normalize="ratio",summaryOperation="mean")
 #   bwToGeneExp(Exons,GeneExpSignals,LibStrand, mc.cores=1,normalize="quantile",
 #    sampleIDs=NULL,summaryOperation="mean")
 #   bwToGeneExp(Exons,GeneExpSignals,LibStrand, mc.cores=1,normalize=NULL,
 #    sampleIDs=NULL,summaryOperation="mean")
 #   bwToGeneExp(Exons=Exons,GeneExpSignals=GeneExpSignals,
 #   LibStrand=LibStrand,sampleIDs=sampleIDs, 
                  mc.cores=1,normalize="quantile",summaryOperation="mean")
}
