% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OverlapRegionsGRadjusted.R
\name{ComplexOverlaps}
\alias{ComplexOverlaps}
\title{Criss-cross overlap of interacting regions}
\usage{
ComplexOverlaps(Reg1, Reg2)
}
\arguments{
\item{Reg1}{- character matrix with at least six columns named
"chr1","start1","end1","chr2","start2","end2" which store 
information about coordinates of two interaction regions in the
genome which we want to comape to two other interacting regions.}

\item{Reg2}{- character matrix with at least six columns named
"chr1","start1","end1","chr2","start2","end2" which store 
information about coordinates of two interaction regions in the
genome which we want to comape to two other interacting regions.}
}
\value{
A dataframe of coordinates of all overlapping regions and
metadata that was stored in the input matrices. Each row corresponds
to one combination of Region1-Region2 overlaps. Column names indicate
whether coordinates and additional info come from Reg1 or Reg2alog.
}
\description{
A function that compares four input regions,eg two interacting regions
from one dataset with two interacting regions from the other dataset
in the criss-cross manner. Thus, the input orientation of the 
interacting regions is not strict: the first pair of the region1
from one dataset will be compared to both; the first pair and the
second pair of region2(and its pair is compared in the opposite 
direction). Two coordinate sets ("chr1","start1","end1",
"chr2","start2","end2") per input dataframe are MUST.
}
\details{
Function performs criss-cross overlap between 2 dataframes 
with coordinates from two interacting regions: "chr1","start1","end1",
"chr2","start2","end2". It creates GRanges objects, and performs
\code{\link[reg2gene]{OverlapRegions}} overlaps between 
Reg1Coord1-Reg2Coord1&Reg1Coord2-Reg2Coord2 and vice-versa
Reg1Coord1-Reg2Coord2&Reg1Coord2-Reg2Coord1.
}
