% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkGInteractions.R
\name{confusionMatrix}
\alias{confusionMatrix}
\title{ConfusionTable statistics for the benchmarked reg2Gene object}
\usage{
confusionMatrix(reg2GeneBench, benchCol = "Bench", prefilterCol = NULL,
  thresholdID = NULL, thresholdValue = 0.05, statistics = "PPV")
}
\arguments{
\item{reg2GeneBench}{GInteractions object with added benchmark 
\code{benchmarkAssociations} and OPTIONAL  \code{filterPreBenchGI} 
metadata. However,prior this analysis, it is advised to reduce the number of
true negatives by including only reg2gene entries that could be potentially
benchmarked.To get that info run \code{filterPreBenchGI} and add filter 
column with metadata from this function to reg2GeneBench GInteractions 
object prior running \code{benchmarkAssociations})}

\item{benchCol}{character (default "Bench"), results of benchmarking 
procedure; eg a column name of the metadata which indicates the column where 
the result of benchmarking procedure is stored. A vector of 0's and 1's is
expected.}

\item{prefilterCol}{character (default NULL), results of prefiltering 
procedure; eg a column name of the metadata which indicates the column where 
the result of prefiltering procedure is stored. A vector of 0's and 1's is
expected.}

\item{thresholdID}{character (def:NULL) name which indicates a column where
statistics of the modelling procedure is stored. This column is filtered such
that everything below predefined threshold is considered to be statistically 
significant association (set to be equal to 0), whereas everything above that
threshold is 0".}

\item{thresholdValue}{numeric (def:0.05) A value of a threshold.}

\item{statistics}{character (def "ConfusionMatrix"). Currentlty, 
"ConfusionMatrix" or "PPV" are options.}
}
\value{
integer vector or a list. If "ConfusionMatrix" is requested then the
output is list with following elements:TP,FP,TN,FN,Specificity, Accuracy, PPV,
NPV,F1. Otherwise only PPV is reported.
}
\description{
A function takes as an benchmarked reg2gene object and outputs
confusion table statistics.
}
\details{
Reports statistics based on the confusion matrix. 
There are four different categories in the confusion matrix: 
TP = (number of) true positive: reg2gene entry that was reported to be 
associated (reported gene-enhancer statistics lower than a predefined
threshold) and was benchmarked. 
FP = (number of) false positive: reg2gene entry that was reported to be
associated (reported gene-enhancer statistics lower than a predefined 
threshold) BUT was not overlapped with benchmark dataset
FN = (number of) false negative: reg2gene entry that was NOT reported to
be associated (reported gene-enhancer statistics NOT lower than a predefined 
threshold) BUT is benchmarked
TN = (number of) true negative: reg2gene entry 
that was NOT reported to be associated (reported gene-enhancer
statistics NOT lower than a predefined threshold) AND is NOT benchmarked.
If no benchmark and no statistically significant data is entered, then 0 is 
reported.

The formulas used in this function are: \deqn{Sensitivity = TP/(TP+FN)} 
\deqn{Specificity = TN/(TN+FP)} 
\deqn{Accuracy = (TP+TN)/(TP+FN+FP+TN)}
\deqn{PPV = TP/(TP+FP)} 
\deqn{NPV = TN/(TN+FN)} 
\deqn{F1 = (2*TP)/((2*TP)+FP+FN)}
}
\examples{
require(GenomicRanges)
require(InteractionSet)
   
   reg2GeneBench <- GInteractions(GRReg1_toy,GRReg1_toy$reg)
   reg2GeneBench$PValue <- seq(0, 1,length.out = length(GRReg1_toy))


confusionMatrix(reg2GeneBench,
                thresholdID = "PValue",
                thresholdValue = 0.05,
                benchCol = "anchor1.Bench1Exp",
                prefilterCol = "anchor1.Filter1Exp",
                statistics = "ConfusionMatrix")
                
confusionMatrix(reg2GeneBench,
                thresholdID = "PValue", 
                thresholdValue = 0.05,
                benchCol = "anchor1.Bench1Exp",
                prefilterCol = "anchor1.Filter1Exp",
                statistics = "PPV")
               
               
     
}
