% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associateReg2Gene.R
\name{corResample}
\alias{corResample}
\title{Correlation with resampling p-values}
\usage{
corResample(mat, scaleData = scaleData, method = "pearson", col = 1,
  B = B)
}
\arguments{
\item{mat}{matrix of response and predictor variables}

\item{method}{pearson or spearman}

\item{col}{column number of response variable}

\item{B}{number of random samples (shuffling) of given column}
}
\description{
private function that calculate correlation of given
column matrix with a column and assigns significance to
correlation coefficents based on resampling
}
\keyword{internal}
