% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkGInteractions.R
\name{forceByname}
\alias{forceByname}
\title{Benchmarks help function - forceByName}
\usage{
forceByname(benchData, reg2Gene)
}
\description{
Forces benchmark input object to use gene coordinates of reg2Gene object if
gene names overlap. It is crucial that benchData & reg2Gene both contain
meta-data named "name" -> where gene names are stored, 
and that anchor2 in both objects corresponds to gene object.
}
\details{
Read description for \code{benchmarkAssociations}
}
\author{
Inga Patarcic
}
\keyword{internal}
