% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotF.R
\name{getTSS}
\alias{getTSS}
\title{My stupid help function to get TSS for lists of enhancer~gene interactions 
sets}
\usage{
getTSS(AllGenes, interactions)
}
\description{
My stupid help function to get TSS for lists of enhancer~gene interactions 
sets
}
\author{
IngaPa
}
\keyword{internal}
