% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associateReg2Gene.R
\name{glmnetResample}
\alias{glmnetResample}
\title{internal function that uses glmnet for prediction}
\usage{
glmnetResample(mat, scaleData = scaleData, col = 1, B = B)
}
\arguments{
\item{mat}{matrix of response and predictor variables}

\item{col}{column number of response variable}

\item{B}{number of random samples (shuffling) of response variables}
}
\description{
private function that uses glmnet to predict fiven column from the rest
and assigns significance to coefficents based on resampling
}
\keyword{internal}
