% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associateReg2Gene.R
\name{grlist2gr}
\alias{grlist2gr}
\title{convert GRangesList to GRanges with associated regulatory regions (internal)}
\usage{
grlist2gr(grlist)
}
\arguments{
\item{grlist}{a GRangesList output from \code{\link{regActivityAroundTSS}}}
}
\description{
The function converts output of \code{\link{regActivityAroundTSS}} (a
GRangesList) to a GRanges object by rearranging the rows and columns so
that output GRanges object mainly contains gene name and position with
an additional column "reg" which contains the regulatory region coordinates
}
\examples{

m=readRDS("data/sample_rawActivityMatrices/Roadmap/H3K27ac/ENSG00000140718_FTO.rds")

m=apply(m, 2,as.numeric) # change to numeric matrix
glmnetResample(scale(m),col=1,B=1000)

}
\author{
Altuna Akalin
}
\keyword{internal}
