% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GRReg1_toy}
\alias{GRReg1_toy}
\title{GRanges example to test various functions from reg2gene}
\format{GRanges object with 9 ranges and 7 metadata columns}
\usage{
GRReg1_toy
}
\description{
GRanges example to test various functions from reg2gene
}
\details{
Genome locations are preselected in such way that all possible outputs of
the  benchmarking procedure is captured. *Colums which correspond to the 
expected outcome of the benchmarking procedure are:Bench1Exp,Filter2Exp,
Bench2Exp,Filter1Exp
Where Bench1 andFilter1 corresponds to the benchmarking of GRReg1_toy with 
GRReg2_toy, whereas Bench2 and Filter2 corresponds to the benchmarking of 
GRReg1_toy with itself.
For example,following object GRReg1_toy[7] should be benchmarked 3 times 
with  GRReg2_toy dataset
}
\keyword{datasets}
