% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GRReg2_toy}
\alias{GRReg2_toy}
\title{Sample file for Region2 imported as data frame}
\format{GRanges object with 14 ranges and 3 metadata columns}
\usage{
GRReg2_toy
}
\description{
ComplexOverlaps() should report 2 overlaps since the 2nd coordinates-pair of Reg2 is a
flipped version of the 1st coordinates-pair of Reg2. ExpectedOverap column indicates what
kind of overlap is expected to observe.
}
\keyword{datasets}
