% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OverlapRegionsGRadjusted.R
\name{OverlapRegions}
\alias{OverlapRegions}
\title{Overlap regions and/or interacting regions and/or their metadata}
\usage{
OverlapRegions(Reg1, Reg2)
}
\arguments{
\item{Reg1}{- character matrix with at least three columns named:
"chr1","start1","end1" which store information about coordinates
of the regions which we want to comape. Additionally,
"chr2","start2","end2" columns or any other column in the
matrix (such as SNP ID, proxy SNP ID,or reported gene).}

\item{Reg2}{- character matrix with at least three columns named:
"chr1","start1","end1" which store information about coordinates
of the regions which we want to compare Additionally,
"chr2","start2","end2" columns or any other column in the
matrix (such as SNP ID, proxy SNP ID,or reported gene).}
}
\value{
A dataframe of coordinates of all overlapping regions.
Each row corresponds to the combination of overlapped rows from 2 input
dataframes based on either: a) simple coordinates overlap, b) coordinates overlap + 
additionaly filtered based on byReg1 and byReg2 defined columns overlap, c) dual 
coordinates overlap between interacting regions. Column names indicate whether 
coordinates and additional info originates from Reg1 or Reg2alog.
}
\description{
The function that compares two input regions or a combinations of two
interacting regions in the genome and reportes regions present in both
input datasets. If one coordinate set per input dataframe is reported
("chr1","start1","end1") then regions from one set, Reg1, is identified
in the other dataset,Reg2, and repoted.
If two coordinate sets ("chr1","start1","end1","chr2","start2","end2")
per input dataframes are reported (eg two interacting regions in test
and benchmark dataset) then both of them are compared,and if both 
coordinates regions overlap they are reported.
Third, an overlap of 2 dataframes can be identified based on the
overlap between genomic coordinates of two regions (regions originated from
two different datasets) and meta-data column (useful for testing overlaps
with eQTL studies based on the gene name).
}
\details{
Function that performs:a) simple overlap between 2 
dataframes of coordinates stored as  "chr1","start1","end1". It creates
GRanges object, and then it runs \code{findOverlaps}.
If input dataframe contains 2 locations of interacting regions:
"chr1","start1","end1","chr2","start2","end2" then it compares:
(Reg1coord1 with Reg2coord1)&(Reg1coord2 with Reg2coord2), and returns rows
where both combinations of overlaps are TRUE. If byReg1 and byReg1 different 
than NULL then overlapped regions are additionally filtered based on the
overlap between byReg1 and byReg1 defined columns
}
