% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg2gene.R
\docType{package}
\name{reg2gene}
\alias{reg2gene}
\alias{reg2gene-package}
\title{reg2gene: An R package for predicting the target genes for regulatory elements}
\description{
reg2gene: An R package for predicting the target genes for regulatory elements
}
