% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataIntegRFunctions.R
\name{regActivity}
\alias{regActivity}
\title{Calculates regulatory activity over pre-defined regions}
\usage{
regActivity(regRegions, activitySignals, sampleIDs = NULL, isCovNA = FALSE,
  summaryOperation = "mean", normalize = NULL, mc.cores = 1)
}
\arguments{
\item{regRegions}{a GRanges object that contains regulatory regions
over which the regulatory activity will be calculated.}

\item{activitySignals}{a named list of BigWig files. Names correspond to 
unique sample ids/names.}

\item{sampleIDs}{NULL (default). A vector of unique sample
ids/names(.bw files), ordered as the bigwig files are ordered. When NULL
basenames of .bw files is used as a unique sample ids/names.}

\item{isCovNA}{(def:FALSE), if this is set to TRUE, uncovered
bases are set to NA, this is important when dealing with methylation
data, where uncovered bases on a bigWig file do not mean zero methylation.}

\item{summaryOperation}{"mean"(default).
This designates which summary operation should be used over the regions.
Currently, only mean is available, but "median" or "sum" will be implemented
in the future.}

\item{normalize}{NULL(default). Optional "quantile" and "ration"
If set to "quantile" activity measures are quantile normalized as 
implemented in \code{\link[preprocessCore]{normalize.quantiles}} and
returned ; if set to "ration" then "median ration method" implemented
as \code{\link[DESeq2]{estimateSizeFactorsForMatrix}} is used to normalize 
activity measures.}

\item{mc.cores}{(def:1) Define the number of cores to use;
at most how many child processes will be run simultaneously 
using mclapply from parallel package. Parallelization requires at 
least two cores.}
}
\value{
a GRanges object where its meta-columns correspond
        to calculated acvitity measures and column names 
        correspond to provided sample ids or names.
}
\description{
The function calculates regulatory activity from histone
modification, DNAse or methylation signals for pre-defined regulatory
regions and returns a GRanges object with regulatory region locations
and their activity over a set of samples.
}
\details{
regulatory activity is measured by averaging logFC for
histone modification ChIP-seq profiles, or DNAse signal, or methylation
per base.Currently, relevant bigWig files are required to calculate activity       
activity. This function might be extended to work with BAM files
in the future. 

#@examples
#      load("/data/akalin/Projects/AAkalin_reg2gene/reg2gene/pkg/inst/extdata/regRegions.RData")
#      activitySignals <- c("pkg/inst/extdata/E085-H3K27ac.chr10.fc.signal.bigwig",
#                           "pkg/inst/extdata/E066-H3K27ac.chr10.fc.signal.bigwig")
#      regActivity(regRegions,activitySignals)
#      sampleIDs <- c("E085","E066")
#      regActivity(regRegions,activitySignals,sampleIDs)
}
