% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataIntegRFunctions.R
\name{regActivityAroundTSS}
\alias{regActivityAroundTSS}
\title{Identify regulatory regions around provided TSSes}
\usage{
regActivityAroundTSS(regActivity, tss, upstream = 5e+05, downstream = 5e+05,
  mc.cores = 1)
}
\arguments{
\item{regActivity}{a GRanges object output from \code{regActivity}
function}

\item{tss}{a GRanges object output from \code{bwToGeneExp} that
contains the TSS location and associated gene expression values 
per cell type or condition as meta data. Each row should have a
"name" and "name2" columns for unique id or name/symbol for the
gene which the TSS is associated with. One is Ensembl id 
and the other could be used for the gene symbol. Other metadata
column names should represent sample names/ids and should
match the GRanges object provided via regActivity argument.}

\item{upstream}{number of basepairs upstream from TSS to look for
regulatory regions. default 500kb}

\item{downstream}{number of basepairs downstream from TSS to look for 
regulatory regions. default 500kb}

\item{mc.cores}{(def:1) Define the number of cores to use;
at most how many child processes will be run simultaneously 
using mclapply from parallel package. Parallelization requires at 
least two cores.}
}
\value{
An output of regActivityAroundTSS() is a GRangesList object that 
contains per gene GRanges with location of the corresponding TSS and
regulatory regions identified around that gene. Names for the GRangesList 
are unique gene ids/names. Metadata for each GRanges object in the 
GRangesList represents regulatory activity and gene expression quantified 
across a number of samples (.bw files) that have matched IDs in RNA-Seq and
CHiP-Seq,DNase-Seq or bisulfite sequencing experiment. 
The GRanges objects have the following metadata columns:
 1. featureType: either "gene" or "regulatory"
 2. name: name/id for gene and enhancers. Gene name could be id from a 
 database enhancer name should be in the format as follows "chr:start-end"
 3. name2: a secondary name for the feature, such as gene symbol "PAX6" etc. 
 not necessary for enhancers could be NA
 4. other columns: numeric values for gene expression or regulatory actvity.
   Column names represent sample names/ids.
}
\description{
The function identifies the regulatory regions around provided
TSSes over a pre-defined window. The function needs a GRanges
object for TSSes with meta-columns corresponding to expression
levels in different cell types or conditions.
}
\details{
only enhancers located within (+/-)upstream/downstream of TSS 
are identified,extracted and reported in output (together with info
about gene expression). Sample id's (corresponding to the cell types 
or conditions) are included only if both, 1) gene expression values and
2) quantified regulatory activity are available in TSS and 
regActivity objects. Non-overlapping cell types are excluded.
}
\examples{
 # INPUT1: CREATING Toy example for quantified gene expression:

geneExpression <- GRReg1_toy
  geneExpression$bw1 <- rep(1,length(GRReg1_toy))
  geneExpression$bw2 <- rep(2,length(GRReg1_toy))
  geneExpression$bw3 <- rep(3,length(GRReg1_toy))
  
  # INPUT2: CREATING TOY example for quantified enhancer activity:
  
regRegion <- GRReg2_toy
   regRegion$bw1 <- rep(3,length(regRegion))
   regRegion$bw2 <- rep(4,length(regRegion))

# Overlapping quantified enhancer activity and gene expression:
# OUTPUT  regActivityAroundTSS():
regActivityAroundTSS(regActivity=regRegion, tss=geneExpression,
upstream=1,downstream=1)

# when different upstream/downstream argument is used:

regActivityAroundTSS(regRegion,geneExpression,upstream=5,downstream=5)


}
