% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associateReg2Gene.R
\name{rfResample}
\alias{rfResample}
\title{internal RandomForest prediction function}
\usage{
rfResample(mat, col = 1, B = 1000, ...)
}
\arguments{
\item{mat}{matrix of response and predictor variables}

\item{col}{column number of response variables}

\item{B}{number of random samples (shuffling) of response variables}

\item{...}{further arguments to randomForest, not implemented yet}
}
\description{
private function that uses randomForest to predict fiven column from the rest
and assigns significance to
correlation coefficents based on resampling
}
\examples{
m=readRDS("data/sample_rawActivityMatrices/Roadmap/H3K27ac/ENSG00000140718_FTO.rds")

m=apply(m, 2,as.numeric) # change to numeric matrix
rfResample(scale(m),col=1,B=1000)

}
\keyword{internal}
