% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voteAssociations.R
\name{voteAssociations}
\alias{voteAssociations}
\title{Majority vote decision for (regulatory region)-gene associations}
\usage{
voteAssociations(associations, qval.cutoff = 0.05, vote.threshold = 0.5)
}
\arguments{
\item{associations}{a named \code{GRangesList} where each element is a set of
associations returned by \code{\link{associateReg2Gene}} or
\code{\link{metaAssociations}}.}

\item{qval.cutoff}{q-value cutoff that will be used to filter elements in the
input \code{GRangesList}. This argument will only take effect when the
\code{GRanges} object has a numeric "qval" column. If the input object
lacks this column, every association in the object will be treated as a valid
association prediction.}

\item{vote.threshold}{A value between 0 and 1, designates the threshold
needed for fraction of votes necessary to retain an association. Defaults to
0.5, meaning fraction of votes should be  greater than or equal to 0.5 to
retain association.}
}
\value{
A \code{GRanges} object that contains votes for associations from
the voting procedure. The object will contain meta-columns for individual
votes and vote stastics. The object can be further filtered to obtain
the desired level of votes using \code{[]} or equivalent methods.
}
\description{
The function calculates regularatory region to gene associations based on
a majority vote. Multiple associations output by different methods and
data sets can be merged this way, and only the associations that have
support in \code{vote.threshold} fraction of the datasets will be retained.
}
\examples{
\dontrun{
#
assoc=system.file("extdata", "sampleAssociationsGrlist.rds",
                      package = "reg2gene")

voted=voteAssociations(assoc, qval.cutoff=0.05,vote.threshold=0.5)

# get all the vote counts
voted=voteAssociations(assoc, qval.cutoff=0.05,vote.threshold=0)
}
}
\author{
Altuna Akalin
}
